improv: styling fixes
This matches the flatpak sources context drawer to newest designs, and adds a padding of `space_m` at the bottom of pages, to prevent content touching the edge of the window.

Also centers the app screenshots, and makes them a bit smaller in height (since it looks better with applets).
The height could exactly match the screenshot height (except when limited by window width), but that causes the content to jump up and down too much, so a fixed height is likely better.